Refactor email endpoint into component Change success from 200 to 201
