+++ app/src/docs/v1.api-spec.yaml
@@ -59,7 +59,7 @@ paths:
             schema:
               $ref: '#/components/schemas/EmailRequest'
       responses:
-        '200':
+        '201':
           description: Responds with the correspondence id(s)? of the message
           content:
             application/json:
@@ -97,7 +97,7 @@ paths:
             schema:
               $ref: '#/components/schemas/EmailMergeRequest'
       responses:
-        '200':
+        '201':
           description: Responds with the correspondence id(s)? of the message
           content:
             application/json:
